fix: circle dep
