+++ packages/core/src/context/componentLoader.ts
@@ -2,11 +2,9 @@ import {
   CONFIGURATION_KEY,
   CONFIGURATION_OBJECT_KEY,
   DecoratorManager,
-  IComponentInfo,
-  InjectionConfigurationOptions,
   MAIN_MODULE_KEY,
 } from '../decorator';
-import { IMidwayGlobalContainer, ScopeEnum } from '../interface';
+import { IComponentInfo, IMidwayGlobalContainer, InjectionConfigurationOptions, ScopeEnum } from '../interface';
 import { FunctionalConfiguration } from '../functional';
 import { MetadataManager } from '../decorator/metadataManager';
 import { MidwayConfigService } from '../service/configService';
