+++ packages/core/src/functional/index.ts
@@ -1,3 +1,2 @@
 export { defineConfiguration, FunctionalConfiguration } from './configuration';
 export * from './hooks';
-export * from './interface';
