--- packages/core/src/functional/interface.ts
@@ -1,4 +1,0 @@
-import { InjectionConfigurationOptions } from '../decorator/common/configuration';
-import { ILifeCycle } from '../interface';
-
-export type FunctionalConfigurationOptions = InjectionConfigurationOptions & ILifeCycle;
